chore: reorganize svg assets by moving logoimage
