+++ src/components/common/Button.tsx
@@ -9,47 +9,47 @@ interface ButtonProps {
 }
 
 export const Button = ({
-  children,
-  variant = 'filled',
-  size = 'xl',
-  disabled = false,
-  onClick,
+	children,
+	variant = 'filled',
+	size = 'xl',
+	disabled = false,
+	onClick,
 }: ButtonProps) => {
-  const baseStyle = `
+	const baseStyle = `
     flex justify-center items-center gap-[10px] rounded-[10px]
     transition-all font-['Pretendard']
   `;
 
-  const sizeStyles = {
-    sm: 'w-[200px] p-[8px] text-[14px]',
-    md: 'w-[300px] p-[10px] text-[16px]',
-    lg: 'w-[340px] p-[12px] text-[18px]',
-    xl: 'w-full p-[16px] text-[18px]', 
-  };
+	const sizeStyles = {
+		sm: 'w-[200px] p-[8px] text-[14px]',
+		md: 'w-[300px] p-[10px] text-[16px]',
+		lg: 'w-[340px] p-[12px] text-[18px]',
+		xl: 'w-full p-[16px] text-[18px]', 
+	};
 
-  const variantStyles = {
-    filled: 'bg-[#3B4599] text-[#F0F2F7] hover:bg-[#2C3374] cursor-pointer',
-    outlined: 'bg-white text-[#3B4599] border-[1.5px] border-[#3B4599] hover:bg-[#F5F6FA] cursor-pointer',
-    hover: 'bg-[#2C3374] text-[#F0F2F7] cursor-pointer',
-  };
+	const variantStyles = {
+		filled: 'bg-[#3B4599] text-[#F0F2F7] hover:bg-[#2C3374] cursor-pointer',
+		outlined: 'bg-white text-[#3B4599] border-[1.5px] border-[#3B4599] hover:bg-[#F5F6FA] cursor-pointer',
+		hover: 'bg-[#2C3374] text-[#F0F2F7] cursor-pointer',
+	};
 
-  const disabledStyle = 'bg-[#A0A7B2] text-white cursor-not-allowed';
+	const disabledStyle = 'bg-[#A0A7B2] text-white cursor-not-allowed';
 
-  const fullClassName = `
+	const fullClassName = `
     ${baseStyle}
     ${sizeStyles[size]}
     ${disabled ? disabledStyle : variantStyles[variant]}
   `;
 
-  return (
-    <button
-      onClick={onClick}
-      disabled={disabled}
-      className={fullClassName}
-    >
-      {children}
-    </button>
-  );
+	return (
+		<button
+			onClick={onClick}
+			disabled={disabled}
+			className={fullClassName}
+		>
+			{children}
+		</button>
+	);
 };
 
 export default Button;
